requirements.txt
# # AI Models
# google-generativeai>=0.3.0
# torch>=2.0.0
# transformers>=4.35.0
# accelerate>=0.24.0
# einops>=0.7.0
# timm>=0.9.0

# # Image Processing
# pillow>=10.0.0
# opencv-python>=4.8.0

# AI Models
openai>=1.0.0

# Data & Config
pandas>=2.0.0
python-dotenv>=1.0.0

# Firebase Cloud Messaging
firebase-admin>=6.0.0

# Benchmarking
matplotlib>=3.7.0

# A2A (Agent-to-Agent) Communication
fastapi>=0.100.0
uvicorn>=0.23.0
httpx>=0.24.0
pydantic>=2.0.0

# RAG (Retrieval-Augmented Generation)
chromadb>=0.4.0
sentence-transformers>=2.2.0
